requirements.txt
django-rest-knox==4.2.0
djangorestframework==3.13.1
mysqlclient==2.0.1
Pillow==9.1.0
drf-spectacular==0.22.0